feat(settings): half-px chat font size steps, 10-20 range
Interface font size becomes f32 with 0.5px steps (Appearance stepper
and Cmd-=/Cmd--/Cmd-0 zoom) and a 10-20 range; code font keeps
whole-px steps and its 24px ceiling. Message text, composer and the
rem base already read the setting — they now follow half-px sizes.

+++ src/chat_collapse.rs
@@ -142,7 +142,7 @@ pub(crate) fn collapse_wrap(mc: MsgCtx, body: AnyElement, state: Collapse, ws: &
             .child(collapse_bar(mc, "Show less", IconName::ChevronUp, ws, cx))
             .into_any_element(),
         Collapse::Collapsed => {
-            let clip = px(f32::from(ws.read(cx).font_size) * LINE_HEIGHT * COLLAPSE_LINES as f32);
+            let clip = px(ws.read(cx).font_size * LINE_HEIGHT * COLLAPSE_LINES as f32);
             // The fade ends on the color behind the text: the accent bubble
             // for a user message, the chat surface for an assistant reply.
             let surface = if mc.msg.role == Role::User { cx.theme().accent } else { cx.theme().background };
